added constants file
