+++ samples/methods/info-browser.js
@@ -2,10 +2,13 @@
 const { LotusRPC } = require('@filecoin-shipyard/lotus-client-rpc')
 const { BrowserProvider: Provider } = require('@filecoin-shipyard/lotus-client-provider-browser')
 const { testnet } = require('@filecoin-shipyard/lotus-client-schema')
-const methods = require('../filecoin/methods')
+const methods = require('../../filecoin/methods')
 const signer = require("@keyko-io/filecoin-signing-tools/js")
+const constants = require("../constants")
 
-const endpointUrl = 'ws://localhost:1234/rpc/v0'
+let endpointUrl = constants.lotus_endpoint
+let tokenPath = constants.token_path 
+
 const provider = new Provider(endpointUrl, {
     token: async () => {
         return "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJBbGxvdyI6WyJyZWFkIiwid3JpdGUiLCJzaWduIiwiYWRtaW4iXX0.JPoCyZQKwHGB2OkXQs17wQW4qKFXtow9-t8d85dm3go"
