+++ samples/methods/info.js
@@ -2,15 +2,18 @@
 const { LotusRPC } = require('@filecoin-shipyard/lotus-client-rpc')
 const { NodejsProvider: Provider } = require('@filecoin-shipyard/lotus-client-provider-nodejs')
 const { testnet } = require('@filecoin-shipyard/lotus-client-schema')
-const hamt = require('../hamt/hamt')
-const methods = require('../filecoin/methods')
+const hamt = require('../../hamt/hamt')
 // const CID = require('cids')
 const fs = require('fs')
+const methods = require('../../filecoin/methods')
+const constants = require("../constants")
 
-const endpointUrl = 'ws://localhost:1234/rpc/v0'
+let endpointUrl = constants.lotus_endpoint
+let tokenPath = constants.token_path 
+
 const provider = new Provider(endpointUrl, {
     token: async () => {
-        return fs.readFileSync('/Users/jpfernandez/.lotus/token')
+        return fs.readFileSync(tokenPath)
     }
 })
 
@@ -21,25 +24,21 @@ async function load(a) {
     return res.Obj
 }
 
-let spec = [
-    ["target", "address"],
-    ["sent", "bigint"],
-    ["method", "int"],
-    ["params", ["cbor", [["verifier", "address"], ["cap", "bigint"]]]],
-    ["signers", ["list", "address"]],
-]
-
+let schema = {
+    type: "hamt",
+    key: "address",
+    value: "bigint",
+}
 
 async function run() {
     while (true) {
         const head = await client.chainHead()
         const state = head.Blocks[0].ParentStateRoot['/']
         console.log("height", head.Height, state)
-        const data = (await client.chainGetNode(`${state}/@Ha:t080/1/6`)).Obj
-        console.log(JSON.stringify(data, null, 2))
-        await hamt.forEach(data, load, function (k, v) {
-            console.log("key", hamt.readVarInt(k) / 2n, methods.decode(spec, v))
-        })
+        const verifiers = (await client.chainGetNode(`${state}/@Ha:t06/1/1`)).Obj
+        let dta = methods.decode(schema, verifiers)
+        console.log(await dta.asObject(load))
+        console.log(await dta.find(load, "t01003"))
         await new Promise(resolve => { setTimeout(resolve, 1000) })
     }
 }
