+++ samples/methods/message.js
