+++ samples/methods/propose-verifier.js
@@ -3,11 +3,14 @@ const {NodejsProvider: Provider} = require('@filecoin-shipyard/lotus-client-prov
 const {testnet} = require('@filecoin-shipyard/lotus-client-schema')
 const fs = require('fs')
 const signer = require("@keyko-io/filecoin-signing-tools/js")
-const methods = require('../filecoin/methods')
+const methods = require('../../filecoin/methods')
+const constants = require("../constants")
 
-const endpointUrl = 'ws://localhost:1234/rpc/v0'
+let endpointUrl = constants.lotus_endpoint
+let tokenPath = constants.token_path 
+
 const provider = new Provider(endpointUrl, {token: async () => {
-    return fs.readFileSync('/Users/jpfernandez/.lotus/token')
+    return fs.readFileSync(tokenPath)
 }})
 
 const client = new LotusRPC(provider, { schema: testnet.fullNode })
@@ -17,7 +20,15 @@ let key = signer.keyDerive(mnemonic, "m/44'/1'/1/0/2", "")
 console.log("address", key.address)
 
 async function main() {
-    await methods.sendTx(client, key, methods.encodeSend("t01000"))
+    let tx = methods.rootkey.propose(methods.verifreg.addVerifier("t01004", 100000000000000000000000000000000000000000n))
+    /*
+    console.log(tx)
+    console.log("here", methods.encodeAddVerifier("t01004", 100000000000000000000000000000000000000000n).params.toString("hex"))
+    let arg = methods.encodePropose("t080", methods.encodeAddVerifier("t01004", 100000000000000000000000000000000000000000n))
+    console.log(arg.params.toString("hex"))
+    console.log(tx.params.toString("hex"))
+    */
+    await methods.sendTx(client, key, tx)
     process.exit(0)
 }
