+++ samples/methods/tx-client.js
@@ -1,7 +1,8 @@
 const signer = require("@keyko-io/filecoin-signing-tools/js")
 const fetch = require('node-fetch')
-const methods = require('../filecoin/methods')
+const methods = require('../../filecoin/methods')
 const cbor = require('cbor')
+const constants = require("../constants")
 
 const mnemonic = 'robot matrix ribbon husband feature attitude noise imitate matrix shaft resist cliff lab now gold menu grocery truth deliver camp about stand consider number'
 let key = signer.keyDerive(mnemonic, "m/44'/1'/1/0/2", "")
